requirements.txt
fastapi
loguru
pyngrok
python-dotenv
vocode
